+++ core/php/AbeilleParser.php
@@ -33,7 +33,7 @@
     include_once __DIR__.'/AbeilleZigateConst.php'; // Zigate constants
     include_once __DIR__.'/AbeilleZigbee.php'; // Zigbee specific functions
     include_once __DIR__.'/AbeilleZigbeeConst.php'; // Zigbee constants
-    include_once __DIR__.'/../class/AbeilleCmd.class.php';
+    // include_once __DIR__.'/../class/AbeilleCmd.class.php';
     include_once __DIR__.'/../class/AbeilleParser.class.php';
     include_once __DIR__.'/../class/Abeille.class.php';
     include_once __DIR__.'/AbeilleOTA.php';